backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI-powered code analysis and chat API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "chromadb>=1.0.16",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "lark>=1.2.2",
    "loguru>=0.7.3",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.13",
    "uvicorn[standard]>=0.35.0",
]
